placing castortoolspy into right place
